requirements.txt
asgiref==3.6.0
Django==4.1.4
django-ninja==0.20.0
django-ninja-apikey==0.1.0
gunicorn==20.1.0
orjson==3.8.3
pydantic==1.10.2
sqlparse==0.4.3
typing_extensions==4.4.0
